made some changes to the amacrine code

+++ code/projects/glm/glm_core/prep_stimcelldependentGPXV.m
@@ -49,10 +49,10 @@ fitmoviestats.span     =  inputstats.range;
 fitmoviestats.normmean =  inputstats.mu_avgIperpix / inputstats.range;
 
 stim   = double(stim);
-stim   = stim / fitmoviestats.span;
+stim   = stim / double(fitmoviestats.span);
 
 if strcmp(GLMType.nullpoint, 'mean')
-    stim = stim - fitmoviestats.normmean;
+    stim = stim - double(fitmoviestats.normmean);
 else
     error('you need to fill in how to account for stimulus with a different nullpoint')
 end
